+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- [TD-2044] Execute rules when I have permissions to execute
+
 ## [3.3.0] 2019-08-05
 
 ### Added